♻️ refactoring : 컴포넌트 도메인별로 정리
