+++ src/components/ads/AdsBlock.tsx
