+++ src/components/event/EventBlock.tsx
