+++ src/components/route/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 // src/components/PrivateRoute.tsx
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '../hooks/useAuth'
+import { useAuth } from '../../hooks/useAuth'
 
 const PrivateRoute = ({ children }: { children: JSX.Element }) => {
   const { isAuthenticated } = useAuth()
