+++ src/components/route/PublicRoute.tsx
@@ -1,6 +1,6 @@
 // src/components/PublicRoute.tsx
 import { Navigate } from 'react-router-dom'
-import { useAuth } from '../hooks/useAuth'
+import { useAuth } from '../../hooks/useAuth'
 
 const PublicRoute = ({ children }: { children: JSX.Element }) => {
   const { isAuthenticated } = useAuth()
